app/requirements.txt
streamlit>=1.38.0
openai>=1.66.0
azure-identity>=1.17.0
azure-search-documents>=11.6.0
azure-storage-blob>=12.20.0
python-dotenv>=1.0.0
